Format total number of days as an integer

+++ TimeFormatters/DaysTimeFormatter.cs
@@ -13,7 +13,7 @@ namespace LiveSplit.TimeFormatters
                 var builder = new StringBuilder();
 
                 if (time.Value.TotalDays >= 1)
-                    builder.Append(time.Value.TotalDays).Append("d ");
+                    builder.Append((int)time.Value.TotalDays).Append("d ");
 
                 if (time.Value.TotalHours >= 1)
                     builder.Append(time.Value.ToString(@"h\:mm\:ss", CultureInfo.InvariantCulture));
